Move files into model folder
