+++ model/student.rb
