+++ model/teacher.rb
